Basic interface to generating queries
I've added what's almost a DSL for building up queries using the <|> operator
for now (which ORs 2 queries). <|> is probably not a great choice, as it's
used by Control.Applicative, but it will do for now.

I've also made queries pure, because I guess given the same input, you're
generating the same query.