Added support for searching for msets.
In Haskell land given a database, and a query one can call enquire to
recieve an IO [Int] value back, which is a list of document IDs.

In C land, we export an iterator API which provides a get function, a
validity check, and a function to move to the next value in the
iterator.