Restructure and document
Now that this is all starting to fall into shape, we are going to need to
start getting a little more organized. I have used a directory structure
that seems to sort of resemble what I've seen from Haskell projects. I've
chosen "Search.Xapian" as our namespace for now, but I don't know if this
is correct. It can be changed later if necessary.

I've exported some stuff from Search.Xapian too, and documented it using
Haddock. While the public API is not at all set in stone yet, I do think
it will help to discuss it through the Haddock pages (better signal to
noise, imo).